Refactor property unit interfaces and models
- Added new `propertyUnit.interface.ts` to define property unit related types and enums.
- Updated `property.interface.ts` to include `IPropertyWithUnitInfo` interface for enhanced property data handling.
- Removed deprecated `unit.interface.ts` to streamline interface management.
- Refactored `property.service.ts` to utilize new unit information retrieval logic.
- Created `propertyUnit.model.ts` to define Mongoose schema for property units with methods for rent adjustment and maintenance handling.
- Updated `propertyUnit.service.ts` to align with new property unit structure and improve request handling.
- Modified constants in `propertyFormMeta.json` to reflect updated unit type labels.
- Adjusted validation schemas to reference new property unit interfaces.
