+++ app/interfaces/index.ts
@@ -1,7 +1,7 @@
+export * from './propertyUnit.interface';
 export * from './property.interface';
 export * from './profile.interface';
 export * from './events.interface';
 export * from './client.interface';
 export * from './utils.interface';
-export * from './unit.interface';
 export * from './user.interface';
