+++ app/models/index.ts
@@ -2,6 +2,6 @@ import User from '@models/user/user.model';
 import Client from '@models/client/client.model';
 import Profile from '@models/profile/profile.model';
 import Property from '@models/property/property.model';
-import PropertyUnit from '@models/property-unit/property-unit.model';
+import PropertyUnit from '@models/property-unit/propertyUnit.model';
 
 export { PropertyUnit, Property, Profile, Client, User };
